\name{OBsMD-package}
\alias{OBsMD-package}
\alias{OBsMD}
\docType{package}
\title{
Objective Bayesian Model Discrimination in Follow-Up Designs
}
\description{
Implements the objective Bayesian methodology proposed in Consonni and Deldossi in order to choose the optimal experiment that better discriminate between competing models.
}
\details{
\tabular{ll}{
Package: \tab OBsMD\cr
Type: \tab Package\cr
Version: \tab 3.0\cr
Date: \tab 2018-06-01\cr
License: \tab GPL version 3 or later\cr
}
The packages allows you to perform the calculations and analyses described
in Consonni and Deldossi paper in TEST (2015), Objective Bayesian model discrimination 
in follow-up experimental designs.
}
\author{
Author: Laura Deldossi and Marta Nai Ruscone based on Daniel Meyer's code.
Maintainer: Marta Nai Ruscone <mnairuscone@liuc.it>
}
\references{
    Consonni and Deldossi, 2015; Box and Meyer, 1986; Box and Meyer, 1993; Meyer, Steinberg and Box, 1996.
}
\keyword{ package }
\examples{
    data(BM86.data)
}
